# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Backend_internal_models.CastVoteRequest:
    properties:
      candidate_id:
        example: 1
        type: integer
    required:
    - candidate_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pufacomputing.ac.id
    name: API Support
    url: https://compsci.president.ac.id/support
  description: Backend API for PUFA Computer Science with Candidates and Votes Management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://compsci.president.ac.id/terms
  title: PUFA Backend API
  version: "1.0"
paths:
...
  /votes:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all votes (requires vote:view permission)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Filter by candidate ID
        in: query
        name: candidate_id
        type: integer
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of votes
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all votes
      tags:
      - Votes
  /votes/{voteID}/delete:
    delete:
      consumes:
      - application/json
      description: Delete a vote by ID (requires vote:delete permission)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Vote ID
        in: path
        name: voteID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Vote deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vote ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a vote
      tags:
      - Votes
  /votes/can-vote:
    get:
      consumes:
      - application/json
      description: Check if the current user is eligible to vote (year must be 2025,
        hasn't voted yet, valid major)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Eligibility status
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check if user can vote
      tags:
      - Votes
  /votes/candidate/{candidateID}:
    get:
      consumes:
      - application/json
      description: Get all votes for a specific candidate (requires vote:view permission)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Candidate ID
        in: path
        name: candidateID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of votes for the candidate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid candidate ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get votes by candidate
      tags:
      - Votes
  /votes/candidate/{candidateID}/count:
    get:
      consumes:
      - application/json
      description: Get the total number of votes for a specific candidate
      parameters:
      - description: Candidate ID
        in: path
        name: candidateID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Vote count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid candidate ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get vote count for candidate
      tags:
      - Votes
  /votes/cast:
    post:
      consumes:
      - application/json
      description: Cast a vote for a candidate. Any user with year 2025 can vote once
        for a candidate from their major. One vote per user, same major required.
      parameters:
      - description: Bearer token (Year 2025 user)
        in: header
        name: Authorization
        required: true
        type: string
      - description: Vote request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Backend_internal_models.CastVoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Vote cast successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request, already voted, not year 2025, or different major
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cast a vote (Year 2025 users only)
      tags:
      - Votes
  /votes/my-vote:
    get:
      consumes:
      - application/json
      description: Get the current user's vote details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User's vote details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Vote not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my vote
      tags:
      - Votes
  /votes/status:
    get:
      consumes:
      - application/json
      description: Check if the current user has already voted
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Vote status
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Check vote status
      tags:
      - Votes
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"